requirements.txt
pandas==2.2.1
psycopg2_binary==2.9.9
python-dotenv==1.0.1